X-axis calibration
- Builds a composite model based on a set of user specified delta lines.
- Initial guess is calculated based on 10-th and 90-th percentiles of
  the distributions.
The phasespace of the model is flat with big amount of narrow minima.
In order to find the best fit, the experimental data are successively
convolved with gaussians with different widths startign from wide to
narrow. The model for the calibration is 3-th order polynomial, which
potentialy can be changed for higher order polynomial. In order to avoid
solving the inverse of the calibration function, the result is tabulated
and interpolated linarly for each bin of the spectrum.