Solves #64 - a bug in `rc2.spectrum.from_delta_lines`
If only `deltas` and `nbins` are provided, the algorithm calculates
boundaries 10% wider than provided lines. This way there are some zeros
before the first delta line and some after the last one. Otherwise
a confolution will lead to artefacts on the opposite side of the
spectrum.

There was a problem when only a single line is provided -- all x-axis
values were equal to the delta line. The problem is fixed. Now when a
single lien is provided the boundaries are by 20% wider on both sides.